Finished Reputation Bar and checked that 'purchasing' sponsors work, which they do.

+++ CSGameProject/Custom Classes/Sponsors.swift
@@ -7,7 +7,7 @@
 
 struct Sponsor {
     let name: String
-    let reputationRequirement: Double
+    let reputationRequirement: Int
     let lengthOfContract: Int
     let rewardMultiplier: Double
 }
